upenwrt.py: add initial working version
upenwrt is a script that will generate openwrt sysupgrade images with all
user-installed packages baked in, on the fly.
It is implemented as a simple HTTP script that serves a script (/get.sh) and a
single API endpoint (/api/get). The script must be executed on the target device
(presumably in the infamous `curl | sh` pattern) and will collect package lists
and other necessary metadata. The API endpoint is supposed to be called by the
script and will return the resulting image in the response body.

The problem is to separate truly user-installed packages from default packages
(which are marked "user" in the opkg database as well), because the default
packages are much more prone to change between revisions than user packages.

The obvious idea would be to compare opkg databases between the squashfs and the
overlay. However, this will not work if the user already runs an image with some
of their packages baked in (or, by extension, if the tool is used twice or more
in a row). The only viable idea known to the author is to get a vanilla openwrt
image of a matching revision and use its package list as the base.

However, it is infeasible to download and keep vanilla images corresponding to
all possible source revisions (as they are, naturally, rotated daily). Instead,
the openwrt source tree is downloaded and cached. Upon each request, the source
tree is checked out at the revision matching the pre-existing firmware and
targetinfo files are generated by calling the buildsystem. "Vanilla" package
lists are then extracted from targetinfo files and subtracted from submitted
package lists.